gen_amqp behaviour (gen_amqp v0.1.0)

Types

amqp_connection/0

-type amqp_connection() ::
    #{mod := module(),
      conn := pid(),
      channel := binary(),
      confirms_handler := undefined | pid() | {atom(), _} | {via, _, _} | fun((_, _) -> any()),
      buffer := none | queue | module(),
      auto_ack_unhandled := boolean(),
      name => atom() | binary(),
      consumer_tags := [consumer_tag()]}.

amqp_msg/0

-type amqp_msg() :: #amqp_msg{props :: term(), payload :: term()}.

args/0

-type args() :: term().

basic_cancel/0

-type basic_cancel() :: #'basic.cancel'{consumer_tag :: term(), nowait :: term()}.

basic_reject/0

-type basic_reject() :: #'basic.reject'{delivery_tag :: term(), requeue :: term()}.

buffer_module/0

-type buffer_module() :: module().

channel_tag/0

-type channel_tag() :: any().

config/0

-type config() ::
    [{auto_ack_unhandled, boolean()} |
     {buffer, none | queue | buffer_module()} |
     {confirms, confirms_handler()} |
     {heartbeat, pos_integer()} |
     {host, inet:ip_address() | string()} |
     {virtual_host, binary()} |
     {username, binary()} |
     {password, binary()} |
     {port, pos_integer()} |
     {channel_max, pos_integer()} |
     {frame_max, pos_integer()} |
     {connection_timeout, pos_integer()} |
     {ssl_options, list()} |
     {auth_mechanisms, list()} |
     {client_properties, list()} |
     {socket_options, list()}].

confirms_fun/0

-type confirms_fun() :: fun((ack | nack, _) -> any()).

confirms_handler/0

-type confirms_handler() :: false | null | true | pid() | confirms_fun() | confirms_module_function().

confirms_module_function/0

-type confirms_module_function() :: {atom(), atom()}.

consumer_tag/0

-type consumer_tag() :: {queue_name(), tag()}.

deliver_payload/0

-type deliver_payload() ::
    {#'basic.deliver'{consumer_tag :: term(),
                      delivery_tag :: term(),
                      redelivered :: term(),
                      exchange :: term(),
                      routing_key :: term()},
     amqp_msg()}.

exchange/0

-type exchange() :: binary().

file/0

-type file() :: list().

message/0

-type message() :: any().

name/0

-type name() :: {local, atom()} | {global, term()} | {via, atom(), term()}.

options/0

-type options() :: [{timeout, stimeout()} | {debug, [sflag()]}].

parsed_config/0

-type parsed_config() ::
    #{auto_ack_unhandled := boolean(),
      buffer := none | queue | buffer_module(),
      confirms := confirms_handler(),
      network :=
          #amqp_params_network{username :: term(),
                               password :: term(),
                               virtual_host :: term(),
                               host :: term(),
                               port :: term(),
                               channel_max :: term(),
                               frame_max :: term(),
                               heartbeat :: term(),
                               connection_timeout :: term(),
                               ssl_options :: term(),
                               auth_mechanisms :: term(),
                               client_properties :: term(),
                               socket_options :: term()},
      mod := module()}.

queue_name/0

-type queue_name() :: binary().

queue_options/0

-type queue_options() ::
    [{expires, pos_integer()} |
     {mode, binary() | list()} |
     {durable, boolean()} |
     {auto_delete, boolean()} |
     {exclusive, boolean()}].

reply_code/0

-type reply_code() :: pos_integer().

reply_text/0

-type reply_text() :: binary().

return_tag/0

-type return_tag() :: {reply_code(), reply_text(), exchange(), routing_key()}.

routing_key/0

-type routing_key() :: binary().

sflag/0

-type sflag() :: trace | log | {logfile, file()} | statistics | debug.

stimeout/0

-type stimeout() :: pos_integer() | infinity | hibernate.

tag/0

-type tag() :: binary() | {binary(), binary()}.

Callbacks

code_change/3

-callback code_change(OldVsn :: term() | {down, term()}, State :: term(), Extra :: term()) ->
               {ok, NewState :: term()} | {error, Reason :: term()}.

handle_basic_cancel/2
(optional)

-callback handle_basic_cancel(ChannelTag :: channel_tag(), State :: term()) ->
                       {noreply, NewState :: term()} |
                       {noreply, NewState :: term(), timeout() | hibernate} |
                       {stop, Reason :: term(), Reply :: term(), NewState :: term()} |
                       {stop, Reason :: term(), NewState :: term()}.

handle_basic_deliver/2

-callback handle_basic_deliver({Tag :: tag(), Message :: message()}, State :: term()) ->
                        {noreply, NewState :: term()} |
                        {noreply, NewState :: term(), timeout() | hibernate} |
                        {stop, Reason :: term(), Reply :: term(), NewState :: term()} |
                        {stop, Reason :: term(), NewState :: term()}.

handle_basic_reject/2
(optional)

-callback handle_basic_reject(RejectTag :: tag(), State :: term()) ->
                       {noreply, NewState :: term()} |
                       {noreply, NewState :: term(), timeout() | hibernate} |
                       {stop, Reason :: term(), Reply :: term(), NewState :: term()} |
                       {stop, Reason :: term(), NewState :: term()}.

handle_basic_return/2
(optional)

-callback handle_basic_return(ReturnTag :: return_tag(), State :: term()) ->
                       {noreply, NewState :: term()} |
                       {noreply, NewState :: term(), timeout() | hibernate} |
                       {stop, Reason :: term(), Reply :: term(), NewState :: term()} |
                       {stop, Reason :: term(), NewState :: term()}.

handle_call/3

-callback handle_call(Request :: term(), From :: {pid(), Tag :: term()}, State :: term()) ->
               {reply, Reply :: term(), NewState :: term()} |
               {reply, Reply :: term(), NewState :: term(), timeout() | hibernate} |
               {noreply, NewState :: term()} |
               {noreply, NewState :: term(), timeout() | hibernate} |
               {stop, Reason :: term(), Reply :: term(), NewState :: term()} |
               {stop, Reason :: term(), NewState :: term()}.

handle_cast/2

-callback handle_cast(Request :: term(), State :: term()) ->
               {noreply, NewState :: term()} |
               {noreply, NewState :: term(), timeout() | hibernate} |
               {stop, Reason :: term(), NewState :: term()}.

handle_connect/1

-callback handle_connect(State :: term()) ->
                  {noreply, NewState :: term()} |
                  {noreply, NewState :: term(), timeout() | hibernate} |
                  {stop, Reason :: term(), Reply :: term(), NewState :: term()} |
                  {stop, Reason :: term(), NewState :: term()}.

handle_continue/2
(optional)

-callback handle_continue(Info :: term(), State :: term()) ->
                   {noreply, NewState :: term()} |
                   {noreply, NewState :: term(), timeout() | hibernate | {continue, term()}} |
                   {stop, Reason :: term(), NewState :: term()}.

handle_disconnect/2

-callback handle_disconnect(Reason :: term(), State :: term()) ->
                     {noreply, NewState :: term()} |
                     {noreply, NewState :: term(), timeout() | hibernate} |
                     {stop, Reason :: term(), Reply :: term(), NewState :: term()} |
                     {stop, Reason :: term(), NewState :: term()}.

handle_error/4
(optional)

-callback handle_error(From :: basic_deliver | rpc,
             Reason :: atom(),
             {Tag :: binary(), Msg :: any()},
             State :: term()) ->
                {noreply, NewState :: term()} |
                {noreply, NewState :: term(), timeout() | hibernate} |
                {stop, Reason :: term(), Reply :: term(), NewState :: term()} |
                {stop, Reason :: term(), NewState :: term()}.

handle_info/2

-callback handle_info(Info :: timeout() | term(), State :: term()) ->
               {noreply, NewState :: term()} |
               {noreply, NewState :: term(), timeout() | hibernate} |
               {stop, Reason :: term(), NewState :: term()}.

handle_queue_deleted/2
(optional)

-callback handle_queue_deleted(Queue :: queue_name(), State :: term()) ->
                        {noreply, NewState :: term()} |
                        {noreply, NewState :: term(), timeout() | hibernate} |
                        {stop, Reason :: term(), Reply :: term(), NewState :: term()} |
                        {stop, Reason :: term(), NewState :: term()}.

handle_rpc/2
(optional)

-callback handle_rpc(Message :: {invalid_term, binary()} | {message(), term()}, State :: term()) ->
              {reply, Reply :: term(), NewState :: term()} |
              {reply, Reply :: term(), NewState :: term(), timeout() | hibernate} |
              {stop, Reason :: term(), Reply :: term(), NewState :: term()} |
              {stop, Reason :: term(), NewState :: term()}.

init/1

-callback init(Args :: config()) ->
        {ok, State :: term()} |
        {ok, State :: term(), timeout() | hibernate | {continue, term()}} |
        {stop, Reason :: term()} |
        ignore.

terminate/2

-callback terminate(Reason :: normal | shutdown | {shutdown, term()} | term(), State :: term()) -> term().

Functions

abcast(X, Y)

abcast(X, Y, Z)

ack(M)

-spec ack(deliver_payload() | basic_cancel() | binary()) -> ok.

acknowledge a message

bind(Queue, Exchange)

-spec bind(Queue :: binary(), Exchange :: binary()) -> ok.

Bind a queue to an exchange with the default routing key.

bind(Queue, Exchange, RoutingKey)

-spec bind(Queue :: binary(), Exchange :: binary(), RoutingKey :: binary()) -> ok.

call(X, Y)

call(X, Y, Z)

cast(X, Y)

code_change(OldVsn, _, Extra)

connect(Opts)

-spec connect(parsed_config()) -> {ok, amqp_connection()} | {error, term()}.

connect(Opts, Pid)

-spec connect(parsed_config(), atom() | pid()) -> {ok, amqp_connection()} | {error, term()}.

connection_pid(_)

direct_exchange(ExchangeName)

-spec direct_exchange(ExchangeName :: binary()) -> ok.

Create a direct exchange

disconnect()

Disconnect from the AMQP server.

enter_loop(X, Y, Z)

enter_loop(X, Y, Z, A)

enter_loop(X, Y, Z, A, B)

fanout_exchange(ExchangeName)

-spec fanout_exchange(ExchangeName :: binary()) -> ok.

Create a fanout exchange

handle_call(Request, From, _)

handle_cast(Request, _)

handle_continue(Request, _)

handle_info(InPayload, _)

init(_)

Connection behavior options:

auto_ack_unhandled: boolean for automatic acknowledgment

buffer: buffering strategy (none, queue, or custom module)

confirms: message confirmation handling

Connection parameters:

heartbeat: heartbeat interval

host: server hostname or IP

virtual_host: RabbitMQ virtual host

username: authentication username

password: authentication password

Network settings:

port: server port number

channel_max: maximum number of channels

frame_max: maximum frame size

connection_timeout: connection timeout

ssl_options: SSL configuration

auth_mechanisms: authentication mechanisms

client_properties: client properties

socket_options: TCP socket options

multi_call(X, Y)

multi_call(X, Y, Z)

multi_call(A, B, C, D)

nack(M)

-spec nack(deliver_payload() | basic_cancel() | binary()) -> ok.

negative acknowledge a message

payload(_)

-spec payload(deliver_payload()) -> binary().

Extract the payload from the message.

publish(Exchange, Payload)

-spec publish(binary(), term()) -> ok | {ok, buffered} | {error, disconnected | buffer_failed}.

Publish a message to an exchange with the default routing key.

publish(Exchange, RoutingKey, Payload)

-spec publish(binary(), binary(), term()) -> ok | {ok, buffered} | {error, disconnected | buffer_failed}.

Publish a message to an exchange with a routing key.

publish_confirmed(Exchange, Payload)

-spec publish_confirmed(binary(), term()) -> ok | {ok, buffered} | {error, disconnected | buffer_failed}.

Publish a message and wait for confirmation.

publish_confirmed(Exchange, RoutingKey, Payload)

-spec publish_confirmed(binary(), binary(), term()) ->
                     ok | {ok, buffered} | {error, disconnected | buffer_failed}.

publish_confirmed_async(Exchange, RoutingKey, PayloadId, Payload)

-spec publish_confirmed_async(Exchange :: binary(),
                        RoutingKey :: binary(),
                        PayloadId :: any(),
                        Payload :: term()) ->
                           ok | {error, disconnected} | {error, no_confirms_handler}.

queue(QueueName)

-spec queue(binary()) -> {ok, binary()} | {error, term()}.

Create a queue with the given name.

queue(QueueName, Opts)

queue_length(Queue)

Get the length of a queue. This number can be off by 1 if the connection is consuming from the queue.

reply(A, B)

rpc_request(Queue, Payload)

-spec rpc_request(binary(), any()) -> ok | {ok, any()} | {error, disconnected}.

Synchronous RPC call to the given queue. These cannot be buffered if the connection is lost.

rpc_response(_, _, Payload)

-spec rpc_response(amqp_connection(), deliver_payload(), any()) -> ok | {error, disconnected}.

Send a response to an RPC as a synchronous call to the given queue. These cannot be buffered if the connection is lost.

start(Module, Config, Args, Options)

-spec start(Module :: module(), Config :: config(), Args :: args(), Options :: options()) ->
         {ok, pid()} | {error, {already_started, pid()}} | {error, term()}.

start(ServerName, Module, Config, Args, Options)

-spec start(ServerName :: name(),
      Module :: module(),
      Config :: config(),
      Args :: args(),
      Options :: options()) ->
         {ok, pid()} | {error, {already_started, pid()}} | {error, term()}.

start_link(Module, Config, Args, Options)

-spec start_link(Module :: module(), Config :: config(), Args :: args(), Options :: options()) ->
              {ok, pid()} | {error, {already_started, pid()}} | {error, term()}.

start_link(ServerName, Module, Config, Args, Options)

-spec start_link(ServerName :: name(),
           Module :: module(),
           Config :: config(),
           Args :: args(),
           Options :: options()) ->
              {ok, pid()} | {error, {already_started, pid()}} | {error, term()}.

subscribe(Q)

Subscribe to a queue with the given name. This will create a new consumer tag for the given queue. The consumer tag will be returned in the reply.

subscribe(Q, Prefetch)

subscribe(Q, Prefetch, ConnPid)

tag(_)

-spec tag(deliver_payload() | basic_cancel() | basic_reject()) -> any().

Extract the delivery tag from the message.

terminate(Reason, _)

topic_exchange(ExchangeName)

-spec topic_exchange(ExchangeName :: binary()) -> ok.

Create a topic exchange

unsubscribe(Q)

-spec unsubscribe(binary()) -> ok.

unsubscribe_all()

-spec unsubscribe_all() -> ok.

wait_for_confirms(_)

wait_for_confirms(_, Timeout)